Move spec sheet manager
